Revert "Rename lib to pyTransition (#35)"
This reverts commit e01ec2396a6dbed291bcdde2c893c4044a07a87b.
