+++ setup.py
@@ -20,22 +20,22 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from setuptools import setup, find_packages
+from setuptools import setup,  find_packages
 from pathlib import Path
 
+# The directory containing this file
+HERE = Path(__file__).parent
+description = (HERE / "README.md").read_text()
 setup(
-    name='pyTransition',
+    name='transition_lib',
     version='0.1.0',    
-    description='A Python package to interact with the Transition API.',
-    long_description=open('README.md').read(),
-    long_description_content_type='text/markdown',
-    author='Transition City',
+    description=description,
     packages=find_packages(),
     install_requires=['requests', 'requests_mock'],
-    python_requires='>=3.6',
+
     classifiers=[
         'License :: OSI Approved :: MIT License',  
-        'Operating System :: OS Independent', 
+        'Operating System :: Linux', 
         'Programming Language :: Python :: 3',
     ],
 )
