+++ tests/test_transition.py
@@ -1,9 +1,10 @@
 import unittest
+from unittest.mock import patch
 import requests
 import requests_mock
 from datetime import time
 
-from ..pyTransition.transition import Transition
+from ..transition_lib.transition import Transition
 
 class TestTransition(unittest.TestCase):
     def setUp(self):
